backend/requirements.txt
fastapi
uvicorn
yt-dlp
openai
pydantic
python-dotenv
duckduckgo-search
websockets
